[Tiered] Work with generics, avoid hangs.
The code was compiling the cfg->method rather than cfg->orig_method,
the former being the method with generics, the latter being the
instantiation.  Without this, we would attempt to compile a generic
method and hit this assert:

       if (!cfg->gshared)
       	   g_assert (!sig->has_type_parameters);

Switched to use the Coop-friendly synchronization primitives,
otherwise, when running the runtime in hybrid or coop mode.  In this
mode, the rejit thread might be waiting on an OS thread, but the GC
suspend machinery would be waiting for this to cooperate, and that
would deadlock.

As Aleksey explained:

   In your stack trace, the GC is trying to cooperatively suspend the
   recompilation thread (which is a cooperative thread because it's a
   "normal" runtime internal thread), but it's blocked on an OS-level
   condition variable, so it never gets to a safepoint.  And so the GC
   is stuck.